refactor: move maximize button to chart renderer
